athlete info task

+++ autos/models.py
@@ -51,5 +51,5 @@ class ChallengeRecord(models.Model):
 
 class AthleteInfo(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    level = models.IntegerField(null=True, blank=True)
+    level = models.PositiveSmallIntegerField(null=True, blank=True)
     goals = models.CharField(max_length=255, default='')
